+++ src/Comparison/ArrayComparison.php
@@ -5,25 +5,25 @@ namespace PhpAbac\Comparison;
 class ArrayComparison
 {
     /**
-     * @param string $serializedHaystack
+     * @param array $haystack
      * @param mixed  $needle
      *
      * @return bool
      */
-    public function isIn($serializedHaystack, $needle)
+    public function isIn($haystack, $needle)
     {
-        return in_array($needle, unserialize($serializedHaystack));
+        return in_array($needle, $haystack);
     }
 
     /**
-     * @param string $serializedHaystack
+     * @param array $haystack
      * @param mixed  $needle
      *
      * @return bool
      */
-    public function isNotIn($serializedHaystack, $needle)
+    public function isNotIn($haystack, $needle)
     {
-        return !$this->isIn($serializedHaystack, $needle);
+        return !$this->isIn($haystack, $needle);
     }
 
     /**